license fix & version bump
The license field is optional but in the case that it is filled out, the license file should be referenced as follows

    license = { file = "LICENSE" } ✅
    license = "LICENSE" ❌

This was brought up in discord and so. More info in the [toml file standards](https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#license) or the  [docs page](https://docs.comfy.org/registry/specifications#license)!

=====

Version bump to 1.0.1 required to publish to the registry.